pyproject.toml
[project]
name = "tennis-origination"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.0.0",
    "matplotlib>=3.10.8",
    "prizepicks-oddsjam>=2.3.0",
    "prizepicks-originations-models>=1.43.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
keyring-provider = "subprocess"

[[tool.uv.index]]
name = "prizepicks"
url = "https://us-east4-python.pkg.dev/dataverse-analytics-kh8ff/us-east4-python/simple"
explicit = true

[tool.uv.sources]
prizepicks-oddsjam = { path = "../prizepicks-oddsjam-python", editable = true }
prizepicks-originations-models = { path = "../prizepicks-originations-models", editable = true }
